feat(api): 新增 Swagger 文件與 API 路由
- 將發送訊息的路由改為 /api/v1/send
- 新增 Swagger 設定以生成 API 文件
- 重構發送訊息的邏輯至 sendController
