+++ package.json
@@ -21,6 +21,8 @@
     "express": "^5.1.0",
     "nodemon": "^3.1.10",
     "pm2": "^6.0.6",
+    "swagger-jsdoc": "^6.2.8",
+    "swagger-ui-express": "^5.0.1",
     "winston": "^3.17.0",
     "ws": "^8.18.2"
   }
